Add reading BinaryGameDefinition

+++ Geten.Runtime/IO/BinaryGameFileHeader.cs
@@ -6,6 +6,6 @@ namespace Geten.Runtime.IO
     public class BinaryGameFileHeader
     {
         public int MagicNumber { get; set; }
-        public int SectionCount { get; set; }
+        public byte SectionCount { get; set; }
     }
 }
